requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.13
psycopg-binary==3.2.13
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
